# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing music projects and ABC notation files",
        "title": "Zupfmanager API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/import/directory": {
            "post": {
                "description": "Import all ABC notation files from a specified directory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "import"
                ],
                "summary": "Import ABC files from directory",
                "parameters": [
                    {
                        "description": "Import directory request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ImportDirectoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ImportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/import/file": {
            "post": {
                "description": "Import a single ABC notation file into the song database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "import"
                ],
                "summary": "Import a single ABC file",
                "parameters": [
                    {
                        "description": "Import file request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ImportFileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ImportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check if the API server is running",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "error": {
                    "type": "string",
                    "example": "validation failed"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid input provided"
                }
            }
        },
        "ImportDirectoryRequest": {
            "type": "object",
            "required": [
                "directory_path"
            ],
            "properties": {
                "directory_path": {
                    "type": "string",
                    "example": "/path/to/songs/"
                }
            }
        },
        "ImportFileRequest": {
            "type": "object",
            "required": [
                "file_path"
            ],
            "properties": {
                "file_path": {
                    "type": "string",
                    "example": "/path/to/song.abc"
                }
            }
        },
        "ImportResponse": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ImportResult"
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "summary": {
                    "$ref": "#/definitions/ImportSummary"
                }
            }
        },
        "ImportResult": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string",
                    "example": "created"
                },
                "changes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "title",
                        "genre"
                    ]
                },
                "error": {
                    "type": "string",
                    "example": "file not found"
                },
                "filename": {
                    "type": "string",
                    "example": "song.abc"
                },
                "title": {
                    "type": "string",
                    "example": "Amazing Grace"
                }
            }
        },
        "ImportSummary": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer",
                    "example": 7
                },
                "errors": {
                    "type": "integer",
                    "example": 0
                },
                "total": {
                    "type": "integer",
                    "example": 10
                },
                "unchanged": {
                    "type": "integer",
                    "example": 1
                },
                "updated": {
                    "type": "integer",
                    "example": 2
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}